requirements.txt
torch>=2.0.0
modelscope>=1.9.5
diffsynth>=0.1.0
opencv-python>=4.5.0
pillow>=9.0.0
numpy>=1.20.0 